% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cebq_base0}
\alias{cebq_base0}
\title{CEBQ Data - Base 0}
\format{
A data frame with 2 rows and 36 variables. Column 1 is the subject ID ("ID"); columns 2-36 are CEBQ variables ("cebq1" ... "cebq35").
}
\usage{
data(cebq_base0)
}
\description{
An example dataset with base-0 CEBQ data for 2 subjects.
}
\keyword{datasets}
