% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_asbi.R
\name{score_asbi}
\alias{score_asbi}
\title{score_asbi: Scored data from the Adaptive Social Behavioral Inventory}
\usage{
score_asbi(
  asbi_data,
  pna_value,
  base_zero = TRUE,
  id,
  session_id,
  extra_scale_cols = c()
)
}
\arguments{
\item{asbi_data}{a data.frame all items for theAdaptive Social Behavioral Inventory following the naming conventions described above}

\item{pna_value}{value used when participant prefers not to answer/elects to skip}

\item{base_zero}{(logical) TRUE indicates the smallest value assigned to a choice is 0. FALSE indicates indicates the smallest value assigned to a choice is 1. Default = TRUE.}

\item{id}{(optional) name of participant ID column in input data. If included, the output dataset will be matched by id, if not included the output dataset will be in the order of the input data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}

\item{session_id}{(optional) name of session ID column in input data if there are multiple observations per participant. If included, the output dataset will be matched by id and session id, if not included the output dataset will be in the order of the input data. Required to get the phenotype dataset if have multiple sessions (raw data merged with scores.)}

\item{extra_scale_cols}{a vector of character strings that begin with 'asbi' but are not scale items. Any columns in asbi_data that begin with 'asbi' but are not scale items must be included here. Default is empty vector.}
}
\value{
A dataset with subscale scores for theAdaptive Social Behavioral Inventory
}
\description{
This function scores the Adaptive Social Behavioral Inventory and provides subscale scores for the following behaviors: Express, Comply, Disrupt
}
\details{
For data to be scored correctly, the data must be prepared according to the following criteria: \cr
\itemize{
 \item{The data must include all individual questionnaire items except for question 13:}
 \item{The columns/variables must match the following naming convention: 'asbi#' or 'asbi_#' where # is the question number (1-30).}
 \item{All questionnaire responses must be a numeric value ranging from 0-2 (base_zero = TRUE) or 1-3 (base_zero = FALSE) where: }
 \itemize{
   \item{For base_zero = TRUE: 0 = Rarely or Never; 1 = Sometimes; 3 = Almost Always}
   \item{For base_zero = TRUE: 1 = Rarely or Never; 2 = Sometimes; 3 = Almost Always}
  }
 \item{Missing values must be coded as NA}
}
\cr
Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the asbi with IDs, with scale from 0-2, prefer not to answer indicated with 2
asbi_score_data <- score_asbi(asbi_data, base_zero = TRUE, id = 'ID', pna_value = 2)


\dontrun{
}


}
\references{
Hogan, Anne E., Keith G. Scott, and Charles R. Bauer. "The Adaptive Social Behavior Inventory (ASBI): A new assessment of social competence in high-risk three-year-olds." Journal of Psychoeducational Assessment 10.3 (1992): 230-239.

Greenfield, Daryl B., et al. "The adaptive social behavior inventory (ASBI): Evaluation with high-risk preschoolers." Journal of Psychoeducational Assessment 15.4 (1997): 322-333.
}
