% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_bevq.R
\name{score_bevq}
\alias{score_bevq}
\title{score_bevq: Scored data from the BEVQ-19}
\usage{
score_bevq(
  bevq_data,
  pna_value,
  base_zero = TRUE,
  id,
  session_id,
  extra_scale_cols = c()
)
}
\arguments{
\item{bevq_data}{a data.frame all items for theBEVQ-19 following the naming conventions described above}

\item{pna_value}{value used when participant prefers not to answer/elects to skip}

\item{base_zero}{(logical) TRUE indicates the smallest value assigned to a choice is 0. FALSE indicates indicates the smallest value assigned to a choice is 1. Default = TRUE.}

\item{id}{(optional) name of participant ID column in input data. If included, the output dataset will be matched by id, if not included the output dataset will be in the order of the input data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}

\item{session_id}{(optional) name of session ID column in input data if there are multiple observations per participant. If included, the output dataset will be matched by id and session id, if not included the output dataset will be in the order of the input data. Required to get the phenotype dataset if have multiple sessions (raw data merged with scores.)}

\item{extra_scale_cols}{a vector of character strings that begin with 'bevq' but are not scale items. Any columns in bevq_data that begin with 'bevq' but are not scale items must be included here. Default is empty vector.}
}
\value{
A dataset with subscale scores for theBEVQ-19
}
\description{
This function scores the BEVQ-19 and provides subscale scores for the following behaviors: Total habitual intake, water intake, sugar sweetend beverage intake.
}
\details{
*Note - alcholic drinks removed for child study.


For data to be scored correctly, the data must be prepared according to the following criteria: \cr
\itemize{
 \item{The data must include all individual questionnaire items except for question 13:}
 \item{The columns/variables must match the following naming convention with each category label as 'bevq_cat_often' or 'bev_cat_amount'. Categories include:}
 \itemize{
   \item{water: 'water'}
   \item{juice: 'juice'}
   \item{sweetend juice drink/beverate: 'sweet'}
   \item{whole milk: 'wholemilk'}
   \item{reduced fat milk: 'reducedmilk'}
   \item{low/fat-free milk: 'lowmilk'}
   \item{flavored milk: 'flavmilk'}
   \item{soft drinks: 'soda'}
   \item{diet soft drinks: 'diet'}
   \item{sweetend tea: 'sweettea'}
   \item{unsweetend tea/coffee (black): 'tea'}
   \item{energy drinks/sports drinks: 'sports drinks'}
   \item{other: 'other'}
   }
 \item{All questionnaire responses must be a numeric value ranging from 0-6 (base_zero = TRUE) or 1-7 (base_zero = FALSE) where: }
 \itemize{
   \item{How often:}
   \itemize{
     \item{For base_zero = TRUE: 0 = Never or less than 1 time per week; 1 = 1 time per week; 2 = 2-3 times per week; 3 = 4-6 times per week; 4 = 1 time per day; 5 = 2+ times per day; 6 = 3+ times per day}
     \item{For base_zero = FALSE: 1 = Never or less than 1 time per week; 2 = 1 time per week; 3 = 2-3 times per week; 4 = 4-6 times per week; 5 = 1 time per day; 6 = 2+ times per day; 7 = 3+ times per day}
   }
   \item{How much/amount:}
   \itemize{
     \item{For base_zero = TRUE: 0 = Less than 6 fl oz (3/4 cup); 1 = 8 fl oz (1 cup); 2 = 12 fl oz (1 1/2 cups); 3 = 16 fl oz (2 cups); 4 = More than 20 fl oz (2 1/2 cups)}
     \item{For base_zero = FALSE: 1 = Less than 6 fl oz (3/4 cup); 2 = 8 fl oz (1 cup); 3 = 12 fl oz (1 1/2 cups); 4 = 16 fl oz (2 cups); 5 = More than 20 fl oz (2 1/2 cups)}
   }
  }
 \item{Missing values must be coded as NA}
}
\cr
Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the bevq with IDs, with scale from 0-2, prefer not to answer indicated with 2
bevq_score_data <- score_bevq(bevq_data, base_zero = TRUE, id = 'ID', pna_value = 2)


\dontrun{
}


}
\references{
Hedrick VE, Savla J, Comber DL, Flack KD, Estabrooks PA, Nsiah-Kumi PA, Ortmeier S, Davy BM. Development of a brief questionnaire to assess habitual beverage intake (BEVQ-15): sugar-sweetened beverages and total beverage energy intake. J Acad Nutr Diet. 2012 Jun;112(6):840-9. doi: 10.1016/j.jand.2012.01.023. PMID: 22709811; PMCID: PMC3379009. (\href{https://pubmed.ncbi.nlm.nih.gov/22709811/}{PubMed})
}
