% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_body_es.R
\name{score_body_es}
\alias{score_body_es}
\title{score_body_es: Scored data from the Body Esteem Scale}
\usage{
score_body_es(
  bes_data,
  pna_value,
  base_zero = TRUE,
  id,
  session_id,
  extra_scale_cols = c()
)
}
\arguments{
\item{bes_data}{a data.frame all items for the Body Esteem Scale following the naming conventions described above}

\item{pna_value}{value used when participant prefers not to answer/elects to skip}

\item{base_zero}{(logical) TRUE indicates the smallest value assigned to a choice is 0. FALSE indicates indicates the smallest value assigned to a choice is 1. Default = TRUE.}

\item{id}{(optional) name of participant ID column in input data. If included, the output dataset will be matched by id, if not included the output dataset will be in the order of the input data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}

\item{session_id}{(optional) name of session ID column in input data if there are multiple observations per participant. If included, the output dataset will be matched by id and session id, if not included the output dataset will be in the order of the input data. Required to get the phenotype dataset if have multiple sessions (raw data merged with scores.)}

\item{extra_scale_cols}{a vector of character strings that begin with 'bes' but are not scale items. Any columns in bes_data that begin with 'bes' but are not scale items must be included here. Default is empty vector.}
}
\value{
A dataset with subscale scores for the Body Esteem Scale
}
\description{
This function scores the Body Esteem Scale
}
\details{
For data to be scored correctly, the data must be prepared according to the following criteria: \cr
\itemize{
 \item{The data must include all individual questionnaire items except for question 13:}
 \item{The columns/variables must match the following naming convention: 'bes#' or 'bes_#' where # is the question number (1-24).}
 \item{All questionnaire responses must be a numeric value ranging from 0-1 (base_zero = TRUE) or 1-2 (base_zero = FALSE) where: }
 \itemize{
   \item{For base_zero = TRUE 0 = No; 1 = Yes}
   \item{For base_zero = TRUE 1 = No; 2 = Yes}
  }
 \item{Missing values must be coded as NA}
}
\cr
Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the bes with IDs, with scale from 0-1, prefer not to answer indicated with 2
bes_score_data <- score_body_es(bes_data, base_zero = TRUE, id = 'ID', pna_value = 2)


\dontrun{
}


}
\references{
Mendelson, B. K., & White, D. R. (1982). Relation between body-esteem and self-esteem of obese and normal children. Perceptual and motor skills, 54(3), 899-905. https://doi.org/10.2466/pms.1982.54.3.899 (\href{https://pubmed.ncbi.nlm.nih.gov/7099901/}{PubMed})
}
