% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_cbq.R
\name{score_cbq}
\alias{score_cbq}
\title{score_cbq: Scored data from the Child Behavior Questionnaire}
\usage{
score_cbq(
  cbq_data,
  base_zero = TRUE,
  id,
  does_not_apply_value = NA,
  extra_scale_cols = c()
)
}
\arguments{
\item{cbq_data}{a data.frame all items for the Child Behavior Questionnaire following the naming conventions described above}

\item{base_zero}{(logical) TRUE indicates the smallest value assigned to a choice is 0. FALSE indicates indicates the smallest value assigned to a choice is 1. Default = TRUE.}

\item{id}{(optional) name of participant ID column in input data. If included, the output dataset will be matched by id, if not included the output dataset will be in the order of the input data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}

\item{does_not_apply_value}{value used to code a cbq response of "does not apply". Default is NA}

\item{extra_scale_cols}{a vector of character strings that begin with 'cbq' but are not scale items. Any columns in cbq_data that begin with 'cbq' but are not scale items must be included here. Default is empty vector.}
}
\value{
A dataset with subscale scores for the Child Behavior Questionnaire
}
\description{
This function scores the Child Behavior Questionnaire and provides subscale scores for the following behaviors: Activity Level, Anger/Frustration, Approach/Positive Anticipation, Attentional Focusing, Discomfort, Falling Reactivity/Soothability, Fear, High Intesity Pleasure, Impulsivity, Inhibitory Control, Low Intensity Pleasure, Perceptual Sensitivity, Sadness, Shyness, Smiling and Laughter. We can also get the Big 3 subcales: Surgency, Negative Affect, and Effortful Control.
}
\details{
For data to be scored correctly, the data must be prepared according to the following criteria: \cr
\itemize{
 \item{The data must include all individual questionnaire items}
 \item{The columns/variables must match the following naming convention: 'cbq#' or 'cbq_#' where # is the question number (1-94)}
 \item{All questionnaire responses must be a numeric value ranging from 0-6 (base_zero = TRUE) or 1-7 (base_zero = FALSE) where: }
 \itemize{
    \item{For base_zero = TRUE: 0 = Extremely Untrue; 1 = Quite Untrue; 2 = Sightly Untrue; 3 = Neither True nor False; 4 = Slightly True; 5 - Quite True; 6 - Extremely True, does_not_apply_value (param) = does not apply }
    \item{For base_zero = FALSE: 1 = Extremely Untrue; 2 = Quite Untrue; 3 = Sightly Untrue; 4 = Neither True nor False; 5 = Slightly True; 6 - Quite True; 7 - Extremely True, does_not_apply_value (param) = does not apply }
  }
 \item{Missing values must be coded as NA}
 \item{Values must not be reversed scored. This script will apply reverse scoring so all levels must be true to the scale described above}
}
\cr
Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the cbq with IDs, when cbq values range from 0-6, and "does not apply" responses are scored as 7
cbq_score_data <- score_cbq(cbq_data, base_zero = TRUE, id = 'ID', does_not_apply_value = 7)

# scoring for the cbq with IDs, when cbq values range from 1-7, and "does not apply" responses are scored as NA
cbq_score_data <- score_cbq(cbq_data, base_zero = FALSE, id = 'ID', does_not_apply_value = NA)

\dontrun{
}


}
\references{
Putnam SP, Rothbart MK. Development of Short and Very Short Forms of the Children’s Behavior Questionnaire. Journal of Personality Assessment. 2006;87(1):102-112. doi:10.1207/s15327752jpa8701_09 (\href{https://pubmed.ncbi.nlm.nih.gov/16856791/}{PubMed})

Rothbart MΚ, Ahadi SA, Hershey KL. Temperament and Social Behavior in Childhood. Merrill-Palmer Quarterly. 1994;40(1):21-39 (\href{https://www.jstor.org/stable/23087906}{jstore})
}
