% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_cebq.R
\name{score_cebq}
\alias{score_cebq}
\title{score_cebq: Scored data from the Children's Eating Behavior Questionnaire}
\usage{
score_cebq(cebq_data, score_base = TRUE, id, extra_scale_cols = c())
}
\arguments{
\item{cebq_data}{a data.frame all items for the Children's Eating Behavior Questionnaire following the naming conventions described above}

\item{score_base}{the smallest value assigned to a choice is 0 (i.e., range 0-3). Default = TRUE.}

\item{id}{(optional) name of participant ID column in bes_data. If included the output dataset will be matched by id, if not included the output dataset will be in the order of bes_data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}

\item{extra_scale_cols}{a vector of character strings that begin with 'cebq' but are not scale items. Any columns in cebq_data that begin with 'cebq' but are not scale items must be included here. Default is empty vector.}
}
\value{
A dataset with subscale scores for the Children's Eating Behavior Questionnaire
}
\description{
This function scores the Children's Eating Behavior Questionnaire and provides subscale scores for the following behaviors: Food Responsiveness, Emotional Overeating, Enjoyment of Food, Desire to Drink, Satiety Responsiveness, Slowness in Eating, Emotional Undereating, and Food Fussiness.
}
\details{
Recent analyses by Manzano et al., (2021) indicate a better-fitting 3 factor solution for school-age children, therefore, the following subscales are also reported: reward-based eating, picky eating, and emotional eating

To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention: 'cebq#' or 'cebq_#' where # is the question number (1-35)
3) All questions must have the numeric value for the ranging from 0-4 (score_base = TRUE) or 1-5 (score_base = FALSE)
4) This script will apply reverse scoring so all levels must be true to the scale described above

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the CEBQ with IDs
cebq_score_data <- score_cebq(cebq_data, score_base, id = 'ID')


}
\references{
Primary References for the Children's Eating Behavior Questionnaire and Scoring:
Wardle, J., Guthrie, C. A., Sanderson, S., & Rapoport, L. (2001). Development of the children’s eating behaviour questionnaire. Journal of Child Psychology and Psychiatry, 42, 963–970. https://doi.org/10.1017/S0021963001007727 (\href{https://pubmed.ncbi.nlm.nih.gov/11693591/}{PubMed})

Alternative 3-factor scoring:
Manzano MA, Strong DR, Kang Sim DE, Rhee KE, Boutelle KN. Psychometric properties of the Child Eating Behavior Questionnaire (CEBQ) in school age children with overweight and obesity: A proposed three‐factor structure. Pediatric Obesity. 2021;16(10):e12795. doi:10.1111/ijpo.12795 (\href{https://pubmed.ncbi.nlm.nih.gov/33945226/}{PubMed})'
}
