% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_cfpq.R
\name{score_cfpq}
\alias{score_cfpq}
\title{score_cfpq: Scored data from the Comprehensive Feeding Practices Questionnaire}
\usage{
score_cfpq(
  cfpq_data,
  pna_value,
  base_zero = TRUE,
  id,
  session_id,
  extra_scale_cols = c()
)
}
\arguments{
\item{cfpq_data}{a data.frame all items for the Comprehensive Feeding Practices Questionnaire following the naming conventions described above}

\item{pna_value}{value used when participant prefers not to answer/elects to skip}

\item{base_zero}{(logical) TRUE indicates the smallest value assigned to a choice is 0. FALSE indicates indicates the smallest value assigned to a choice is 1. Default = TRUE.}

\item{id}{(optional) name of participant ID column in input data. If included, the output dataset will be matched by id, if not included the output dataset will be in the order of the input data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}

\item{session_id}{(optional) name of session ID column in input data if there are multiple observations per participant. If included, the output dataset will be matched by id and session id, if not included the output dataset will be in the order of the input data. Required to get the phenotype dataset if have multiple sessions (raw data merged with scores.)}

\item{extra_scale_cols}{a vector of character strings that begin with 'cfpq' but are not scale items. Any columns in cfpq_data that begin with 'cfpq' but are not scale items must be included here. Default is empty vector.}
}
\value{
A dataset with subscale scores for the Comprehensive Feeding Practices Questionnaire
}
\description{
This function scores the Comprehensive Feeding Practices Questionnaire and provides subscale scores for: child control, emotional regulation, Encourage balance and variety, environment, food as reward, involvement, modeling, monitoring, pressure, restriction for health, restriction for weight control, and teaching about nutrition
}
\details{
#' For data to be scored correctly, the data must be prepared according to the following criteria: \cr
\itemize{
 \item{The data must include all individual questionnaire items}
 \item{The columns/variables must match the following naming convention: 'cfpq#' or 'cfpq_#' where # is the question number (1-49)}
 \item{All questionnaire responses must be a numeric value ranging from 0-4 (base_zero = TRUE) or 1-5 (base_zero = FALSE) where: }
 \itemize{
    \item{For base_zero = TRUE: 0 = never/disagree; 1 = rarely/slightly disagree; 2 = sometimes/neutral; 3 = often/slightly agree; 4 = always/agree}
    \item{For base_zero = FALSE: 1 = never/disagree; 2 = rarely/slightly disagree; 3 = sometimes/neutral; 4 = often/slightly agree; 5 = always/agree}
  }
 \item{Missing values must be coded as NA}
 \item{Values must not be reversed scored. This script will apply reverse scoring so all levels must be true to the scale described above}
}
Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the CFPQ with IDs, when values range from 0-4
cfpq_score_data <- score_cfpq(cfpq_data, base_zero = TRUE, id = 'ID')

scoring for the CFPQ with IDs, when values range from 1-5
cfpq_score_data <- score_cfpq(cfpq_data, base_zero = FALSE, id = 'ID')

\dontrun{
}

}
\references{
Primary Reference for the Questionnaire and Scoring:
Musher-Eizenman, D., Holub. S., Comprehensive Feeding Practices Questionnaire: Validation of a New Measure of Parental Feeding Practices, Journal of Pediatric Psychology, 32, 8, September 2007, Pages 960–972, https://doi.org/10.1093/jpepsy/jsm037
}
