% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_cfq.R
\name{score_cfq}
\alias{score_cfq}
\title{score_cfq: Scored data from the Child Feeding Questionnaire}
\usage{
score_cfq(cfq_data, score_base = TRUE, restriction_split = FALSE, id)
}
\arguments{
\item{cfq_data}{a data.frame all items for the Child Feeding Questionnaire following the naming conventions described above}

\item{score_base}{the smallest value assigned to a choice is 0 (i.e., range 0-3). Default = TRUE.}

\item{restriction_split}{a boolean incating if the Restriction subscale should be split to remove food as reward items. Default = FALSE. The standard Restriction subscale will always be available. If restriction_split = TRUE, then two adtional scales will be computed: 1) cfq_rest_noreward: questions 17-20, 23-24 and 1) cfq_foodreward: questions 21-22}

\item{id}{(optional) name of participant ID column in bes_data. If included the output dataset will be matched by id, if not included the output dataset will be in the order of bes_data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}
}
\value{
A dataset with subscale scores for the Child Feeding Questionnaire
}
\description{
This function scores the Child Feeding Questionnaire and provides subscale scores for the following behaviors: Perceived Responsibility, Perceived Child Weight, Perceived Parent Weight, Child Weight Concerns, Restriction, Pressure to Eat, and Monitoring
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention: 'cfq#' where # is the question number (1-31). Question 13 be skipped for subscale - Perceived Child Weight due to age range.
3) All questions must have the numeric value for the choice with the base value being either 0 (score_base = TRUE) or 1 (score_base = FALSE)

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the cfq with IDs
cfq_score_data <- score_cfq(cfq_data, score_base = TRUE, id = 'ID')

# scoring for the cfq with extra Restriction subscales
cfq_score_data <- score_cfq(cfq_data, restriction_split = TRUE, id = 'ID')

\dontrun{
}


}
\references{
Birch, L. L., Fisher, J. O., Grimm-Thomas, K., Markey, C. N., Sawyer, R., & Johnson, S. L. (2001). Confirmatory factor analysis of the Child Feeng Questionnaire: A measure of parental attitudes, beliefs and practices about child feeng and obesity proneness. Appetite, 36(3), 201–210. https://doi.org/10.1006/appe.2001.0398 (\href{https://pubmed.ncbi.nlm.nih.gov/11358344/}{PubMed})
}
