% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_cfsq.R
\name{score_cfsq}
\alias{score_cfsq}
\title{score_cfsq: Scored data from the Caregiver's Feeding Style Questionnaire}
\usage{
score_cfsq(
  cfsq_data,
  pna_value,
  base_zero = TRUE,
  id,
  session_id,
  extra_scale_cols = c()
)
}
\arguments{
\item{cfsq_data}{a data.frame all items for the Caregiver's Feeding Style Questionnaire following the naming conventions described above}

\item{pna_value}{value used when participant prefers not to answer/elects to skip}

\item{base_zero}{(logical) TRUE indicates the smallest value assigned to a choice is 0. FALSE indicates indicates the smallest value assigned to a choice is 1. Default = TRUE.}

\item{id}{(optional) name of participant ID column in input data. If included, the output dataset will be matched by id, if not included the output dataset will be in the order of the input data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}

\item{session_id}{(optional) name of session ID column in input data if there are multiple observations per participant. If included, the output dataset will be matched by id and session id, if not included the output dataset will be in the order of the input data. Required to get the phenotype dataset if have multiple sessions (raw data merged with scores.)}

\item{extra_scale_cols}{a vector of character strings that begin with 'cfsq' but are not scale items. Any columns in cfsq_data that begin with 'cfsq' but are not scale items must be included here. Default is empty vector.}
}
\value{
A dataset with subscale scores for the Caregiver's Feeding Style Questionnaire
}
\description{
This function scores the Caregiver's Feeding Style Questionnaire and provides subscale scores for the following behaviors: Demandingness and Responsiveness
}
\details{
For data to be scored correctly, the data must be prepared according to the following criteria: \cr
\itemize{
 \item{The data must include all individual questionnaire items except for question 13:}
 \item{The columns/variables must match the following naming convention: 'cfsq#' or 'cfsq_#' where # is the question number (1-19).}
 \item{All questionnaire responses must be a numeric value ranging from 0-4 (base_zero = TRUE) or 1-5 (base_zero = FALSE) where: }
 \itemize{
   \item{For base_zero = TRUE: 0 = Never; 1 = Rarely; 2 = Sometimes; 3 = Often; 4 = Always}
   \item{For base_zero = TRUE: 1 = Never; 2 = Rarely; 3 = Sometimes; 4 = Often; 5 = Always}
  }
 \item{Missing values must be coded as NA}
}
\cr
Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the cfsq with IDs, with scale from 0-1, prefer not to answer indicated with 2
cfsq_score_data <- score_cfsq(cfsq_data, base_zero = TRUE, id = 'ID', pna_value = 2)


\dontrun{
}


}
\references{
Hughes SO, Power TG, Orlet Fisher J, Mueller S, Nicklas TA. Revisiting a neglected construct: parenting styles in a child-feeding context. Appetite. 2005 Feb;44(1):83-92. doi: 10.1016/j.appet.2004.08.007. Epub 2004 Nov 13. PMID: 15604035.(\href{https://pubmed.ncbi.nlm.nih.gov/15604035/}{PubMed})
}
