% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_cshq.R
\name{score_cshq}
\alias{score_cshq}
\title{score_cshq: Scored data from the Children Sleep Habits Questionnaire}
\usage{
score_cshq(cshq_data, score_base = TRUE, id, reverse_score = FALSE)
}
\arguments{
\item{cshq_data}{a data.frame all items for the Children Sleep Habits Questionnaire following the naming conventions described above}

\item{score_base}{the smallest value assigned to a choice is 0 (i.e., range 0-3). Default = TRUE.}

\item{id}{(optional) name of participant ID column in bes_data. If included the output dataset will be matched by id, if not included the output dataset will be in the order of bes_data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}

\item{reverse_score}{is data already reversed scored (default = FALSE)}
}
\value{
A dataset with total and subscale scores for the Children Sleep Habits Questionnaire
}
\description{
This function scores the Children Sleep Habits Questionnaire and provides both total scores and subscale scores for the following behaviors: Morning Wake Up, Bedtime Resistance, Sleep Onset Delay, Sleep Duration, Sleep Anxiety, Night Wakings, Parasomnias, Sleep Disordered Breathing, and Daytime Sleepiness
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention: 'cshq#' or 'cshq_#' where # is the question number (1-33)
3a) If score_base = FALSE, all questions must have the numeric values 1-3: (3 - Usually, 2 - Sometimes, 1 - Rarely or 1 - not sleepy, 2 - very sleepy, 3 - falls asleep)
3b) If score_base = TRUE, all questions must have the numeric values 0-2: (2 - Usually, 1 - Sometimes, 0 - Rarely or 0 - not sleepy, 1 - very sleepy, 2 - falls asleep)
4) This script will apply reverse scoring so all levels must be true to the scale described above

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

cshq_score_data <- score_cshq(cshq_data, id = 'ID')

\dontrun{
}



}
\references{
Full Measure:
Owens, J. A., Spirito, A., & McGuinn, M. (2000). The Children’s Sleep Habits Questionnaire (CSHQ): Psychometric Properties of A Survey Instrument for School-Aged Children. SLEEP, 23(8), 1043–1052. (\href{https://pubmed.ncbi.nlm.nih.gov/11145319/}{PubMed})
}
