% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_debq.R
\name{score_debq}
\alias{score_debq}
\title{score_debq: INCOMPLETE FUNCTION. Scored data from the Dutch Eating Behavior Questionnaire}
\usage{
score_debq(debq_data, score_base = TRUE, id)
}
\arguments{
\item{debq_data}{a data.frame all items for the Dutch Eating Behavior Questionnaire following the naming conventions described above}

\item{score_base}{the smallest value assigned to a choice is 0 (i.e., range 0-3). Default = TRUE.}

\item{id}{(optional) name of participant ID column in bes_data. If included the output dataset will be matched by id, if not included the output dataset will be in the order of bes_data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}
}
\value{
A dataset with subscale scores for the Dutch Eating Behavior Questionnaire
}
\description{
This function scores the Dutch Eating Behavior Questionnaire and provides subscale scores for the following behaviors:
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention: 'debq#' where # is the question number (1-24).
3) Questions must have the numeric value for the choices: 1, Never | 2, Seldom | 3, Sometimes | 4, Often | 5, Very Often

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the debq with IDs
debq_score_data <- score_debq(debq_data, id = 'ID')


\dontrun{
}


}
\references{
need to add
}
