% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_ecsi2.R
\name{score_ecsi2}
\alias{score_ecsi2}
\title{score_ecsi2: Scored data from the Satter Eating Competence Inventory 2.0}
\usage{
score_ecsi2(
  ecsi_data,
  pna_value,
  base_zero = TRUE,
  id,
  session_id,
  extra_scale_cols = c()
)
}
\arguments{
\item{ecsi_data}{a data.frame all items for the Satter Eating Competence Inventory 2.0 following the naming conventions described above}

\item{pna_value}{value used when participant prefers not to answer/elects to skip}

\item{base_zero}{(logical) TRUE indicates the smallest value assigned to a choice is 0. FALSE indicates indicates the smallest value assigned to a choice is 1. Default = TRUE.}

\item{id}{(optional) name of participant ID column in input data. If included, the output dataset will be matched by id, if not included the output dataset will be in the order of the input data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}

\item{session_id}{(optional) name of session ID column in input data if there are multiple observations per participant. If included, the output dataset will be matched by id and session id, if not included the output dataset will be in the order of the input data. Required to get the phenotype dataset if have multiple sessions (raw data merged with scores.)}

\item{extra_scale_cols}{a vector of character strings that begin with 'ecsi' but are not scale items. Any columns in ecsi_data that begin with 'ecsi' but are not scale items must be included here. Default is empty vector.}
}
\value{
A dataset with subscale scores for the Satter Eating Competence Inventory 2.0
}
\description{
This function scores the Satter Eating Competence Inventory 2.0 and provides subscale scores for the following behaviors: Eating Attitudes, Food Acceptance, Internal Regulation, and Contextual Skills
}
\details{
For data to be scored correctly, the data must be prepared according to the following criteria: \cr
\itemize{
 \item{The data must include all individual questionnaire items except for question 13:}
 \item{The columns/variables must match the following naming convention: 'ecsi#' or 'ecsi_#' where # is the question number (1-16).}
 \item{All questionnaire responses must be a numeric value ranging from 0-4 (base_zero = TRUE) or 1-5 (base_zero = FALSE) where: }
 \itemize{
   \item{For base_zero = TRUE: 0 = Never; 1 = Rarely; 2 = Sometimes; 3 = Often; 4 = Always}
   \item{For base_zero = TRUE: 1 = Never; 2 = Rarely; 3 = Sometimes; 4 = Often; 5 = Always}
  }
 \item{Missing values must be coded as NA}
}
\cr
Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the ecsi with IDs, with scale from 0-1, prefer not to answer indicated with 2
ecsi_score_data <- score_ecsi2(ecsi_data, base_zero = TRUE, id = 'ID', pna_value = 2)


\dontrun{
}


}
\references{
Lohse B, Satter E, Horacek T, Gebreselassie T, Oakland MJ. Measuring eating competence: psychometric properties and validity of the ecSatter Inventory. J Nutr Educ Behav. 2007 Sep-Oct;39(5 Suppl):S154-66. doi: 10.1016/j.jneb.2007.04.371. PMID: 17826696.(\href{https://pubmed.ncbi.nlm.nih.gov/17826696/}{PubMed})

Godleski S, Lohse B, Krall JS. Satter Eating Competence Inventory Subscale Restructure After Confirmatory Factor Analysis. J Nutr Educ Behav. 2019 Sep;51(8):1003-1010. doi: 10.1016/j.jneb.2019.05.287. Epub 2019 Jul 24. PMID: 31350197. (\href{https://pubmed.ncbi.nlm.nih.gov/31350197/}{PubMed})
}
