% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_efcr.R
\name{score_efcr}
\alias{score_efcr}
\title{score_efcr: Scored data from the External Food Cue Responsiveness Scale}
\usage{
score_efcr(efcr_data, base_zero = TRUE, id, session_id, extra_scale_cols = c())
}
\arguments{
\item{efcr_data}{a data.frame all items for the External Food Cue Responsiveness Scale following the naming conventions described above}

\item{base_zero}{(logical) TRUE indicates the smallest value assigned to a choice is 0. FALSE indicates indicates the smallest value assigned to a choice is 1. Default = TRUE.}

\item{id}{(optional) name of participant ID column in input data. If included, the output dataset will be matched by id, if not included the output dataset will be in the order of the input data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}

\item{session_id}{(optional) name of session ID column in input data if there are multiple observations per participant. If included, the output dataset will be matched by id and session id, if not included the output dataset will be in the order of the input data. Required to get the phenotype dataset if have multiple sessions (raw data merged with scores.)}

\item{extra_scale_cols}{a vector of character strings that begin with 'efcr' but are not scale items. Any columns in efcr_data that begin with 'efcr' but are not scale items must be included here. Default is empty vector.}
}
\value{
If 'id' argument is used, returns a list with 2 dataframes: (1) bids_phenotype (contains input efcr_data [values identical to input, underscores removed from efcr items col names, if they existed] and External Food Cue Responsiveness Scale score) and (2) score_dat (contains efcr subscale scores only). If 'id' argument is not used, returns a list with score_dat dataframe only.
}
\description{
This function scores the External Food Cue Responsiveness Scale
}
\details{
For data to be scored correctly, the data must be prepared according to the following criteria: \cr
\itemize{
 \item{The data must include all individual questionnaire items}
 \item{The columns/variables must match the following naming convention: 'efcr#' or 'efcr_#' where # is the question number (1-9)}
 \item{All questionnaire responses must be a numeric value ranging from 0-4 (base_zero = TRUE) or 1-5 (base_zero = FALSE) where: }
 \itemize{
    \item{For base_zero = TRUE: 0 = Never; 1 = Rarely; 2 = Sometimes; 3 = Often; 4 = Always}
    \item{For base_zero = FALSE: 1 = Never; 2 = Rarely; 3 = Sometimes; 4 = Often; 5 = Always}
  }
 \item{Missing values must be coded as NA}
}
\cr
Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the EFCR Scale with IDs
efcr_score_data <- score_efcr(efcr_data, base_zero = TRUE, id = 'ID')


}
\references{
Primary References for the External Food Cue Responsiveness Scale and Scoring:
Masterson TD, Gilbert-Diamond D, Lansigan RK, Kim SJ, Schiffelbein JE, Emond JA. Measurement of external food cue responsiveness in preschool-age children: Preliminary evidence for the use of the external food cue responsiveness scale. Appetite. 2019;139:119-126. doi:10.1016/j.appet.2019.04.024 (\href{https://pubmed.ncbi.nlm.nih.gov/31047939/}{PubMed})

Pollack CC, Emond JA, Masterson TD. Associations between adolescent and young adult External Food Cue Responsiveness (EFCR) and brand recall, product craving and product purchasing in the livestreaming food marketing environment. Public Health Nutr. 2022;25(11):3036-3043. doi:10.1017/S1368980022001628 (\href{https://pubmed.ncbi.nlm.nih.gov/35920082/}{PubMed})
}
