% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_ffbs.R
\name{score_ffbs}
\alias{score_ffbs}
\title{score_ffbs: Scored data from the Family Food Behavior Survey}
\usage{
score_ffbs(
  ffbs_data,
  pna_value,
  base_zero = TRUE,
  id,
  session_id,
  extra_scale_cols = c()
)
}
\arguments{
\item{ffbs_data}{a data.frame all items for the Family Food Behavior Survey following the naming conventions described above}

\item{pna_value}{value used when participant prefers not to answer/elects to skip}

\item{base_zero}{(logical) TRUE indicates the smallest value assigned to a choice is 0. FALSE indicates indicates the smallest value assigned to a choice is 1. Default = TRUE.}

\item{id}{(optional) name of participant ID column in input data. If included, the output dataset will be matched by id, if not included the output dataset will be in the order of the input data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}

\item{session_id}{(optional) name of session ID column in input data if there are multiple observations per participant. If included, the output dataset will be matched by id and session id, if not included the output dataset will be in the order of the input data. Required to get the phenotype dataset if have multiple sessions (raw data merged with scores.)}

\item{extra_scale_cols}{a vector of character strings that begin with 'ffbs' but are not scale items. Any columns in ffbs_data that begin with 'ffbs' but are not scale items must be included here. Default is empty vector.}
}
\value{
A dataset with subscale scores for the Family Food Behavior Survey
}
\description{
This function scores the Family Food Behavior Survey and provides subscale scores for the following behaviors: Maternal Control, Maternal Presences, Child Choice, and Organization
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention: 'ffbs#' or 'ffbs_#' where # is the question number (1-20)
3) All questions must have the numeric value for the choice: 0 - Never True, 1 - Rarely True, 2 - Sometimes, 3 - Often True, 4 - Always True

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the ffbs with IDs
ffbs_score_data <- score_ffbs(ffbs_data, id = 'ID', base_zero = TRUE)

\dontrun{
}


}
\references{
Baughcum, A. E., Powers, S. W., Johnson, S. B., Chamberlin, L. A., Deeks, C. M., Jain, A., & Whitaker, R. C. (2001). Maternal Feeding Practices and Beliefs and Their Relationships to Overweight in Early Childhood: Journal of Developmental & Behavioral Pediatrics, 22(6), 391–408. https://doi.org/10.1097/00004703-200112000-00007 (\href{https://pubmed.ncbi.nlm.nih.gov/11773804/}{PubMed})

McCurdy, K., & Gorman, K. S. (2010). Measuring family food environments in diverse families with young children. Appetite, 54(3), 615–618. https://doi.org/10.1016/j.appet.2010.03.004 (\href{https://pubmed.ncbi.nlm.nih.gov/20227449/}{PubMed})
}
