% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_ffq_helix.R
\name{score_ffq_helix}
\alias{score_ffq_helix}
\title{score_ffq_helix: Scored data from the HELIX study Food Frequency Questionnaire}
\usage{
score_ffq_helix(ffq_data, base_zero = TRUE, id)
}
\arguments{
\item{ffq_data}{a data.frame all items for the HELIX Food Fequency Questionnaire following the naming conventions described above}

\item{base_zero}{(logical) TRUE indicates the smallest value assigned to a choice is 0. FALSE indicates indicates the smallest value assigned to a choice is 1. Default = TRUE.}

\item{id}{(optional) name of participant ID column in input data. If included, the output dataset will be matched by id, if not included the output dataset will be in the order of the input data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}
}
\value{
A dataset with servings per week calculated for the various food categories
}
\description{
This function scores the Food Frequency Questionnaire developed for the HELIX study. Daily frequency of food intake will be calculated according to the following prior scoring conventions by Stratakis et al., 2022:
}
\details{
1) 16 categories: dairy, eggs, meat/meat products, fish/seafood, nuts, vegetables, lentils/pulses, fruit, potatoes, breads/cereals, sweets, beverages, sweet bakery products, salty snacks (chips), added fats, and dressings
2) Ultra-processed foods: cookies, pastries, sugar-sweetened, low-sugar and artificially sweetened beverages, cold meat cuts; ham, dairy desserts, sugar-sweetened and other breakfast cereals, crispbread and rusks; chocolate, sweets, margarine, dressings, and salty snack

*Note: there are also scoring conventions for 11 categories (Lau et al., 2018; dairy, meat/meat products, fish/seafood, vegetables, fruit, potatoes, breads/cereals, sweets, beverages, sweet bakery products, added fats) and 7 categories (Papadopoulou et al., 2019; animal origin--dairy, meat/meat products, fish/seafood and plant origin--vegetables, fruit, potatoes, breads/cereals, and pulses)
3) 

To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The columns/variables must match the following naming convention: 'ffq_xxx#' xxx is the food category and the # is the item number for the category
3) All questions must have the numeric value for the choice with the base value being either 0 (base_zero = TRUE) or 1 (base_zero = FALSE)

Note, as long as variable names match those listed in this script, the dataset can include other variables
}
\examples{

# scoring for the bes with IDs
ffq_score_data <- score_ffq_helix(ffq_data, base_zero = TRUE, id = 'ID')

\dontrun{
} 


}
\references{
Primary for HELIX study: 
Maitre L, De Bont J, Casas M, et al. Human Early Life Exposome (HELIX) study: a European population-based exposome cohort. BMJ Open. 2018;8(9):e021311. doi:10.1136/bmjopen-2017-021311 (\href{https://pubmed.ncbi.nlm.nih.gov/30206078/}{PubMed})

Scoring:
Stratakis N, Siskos AP, Papadopoulou E, et al. Urinary metabolic biomarkers of diet quality in European children are associated with metabolic health. eLife. 2022;11:e71332. doi:10.7554/eLife.71332 (\href{https://pubmed.ncbi.nlm.nih.gov/35076016/}{PubMed})

Lau CHE, Siskos AP, Maitre L, et al. Determinants of the urinary and serum metabolome in children from six European populations. BMC Med. 2018;16(1):202. doi:10.1186/s12916-018-1190-8 (\href{https://pubmed.ncbi.nlm.nih.gov/30404627/}{PubMed})

Papadopoulou E, Haug LS, Sakhi AK, et al. Diet as a Source of Exposure to Environmental Contaminants for Pregnant Women and Children from Six European Countries. Environ Health Perspect. 2019;127(10):107005. doi:10.1289/EHP5324 (\href{https://pubmed.ncbi.nlm.nih.gov/31617753/}{PubMed})
}
