% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_fmcb.R
\name{score_fmcb}
\alias{score_fmcb}
\title{score_fmcb: Scored data from the Feeding to Manage Child Behavior Questionnaire}
\usage{
score_fmcb(
  fmcb_data,
  pna_value,
  base_zero = TRUE,
  id,
  session_id,
  extra_scale_cols = c()
)
}
\arguments{
\item{fmcb_data}{a data.frame all items for the Feeding to Manage Child Behavior Questionnaire following the naming conventions described above}

\item{pna_value}{value used when participant prefers not to answer/elects to skip}

\item{base_zero}{(logical) TRUE indicates the smallest value assigned to a choice is 0. FALSE indicates indicates the smallest value assigned to a choice is 1. Default = TRUE.}

\item{id}{(optional) name of participant ID column in input data. If included, the output dataset will be matched by id, if not included the output dataset will be in the order of the input data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}

\item{session_id}{(optional) name of session ID column in input data if there are multiple observations per participant. If included, the output dataset will be matched by id and session id, if not included the output dataset will be in the order of the input data. Required to get the phenotype dataset if have multiple sessions (raw data merged with scores.)}

\item{extra_scale_cols}{a vector of character strings that begin with 'fmcb' but are not scale items. Any columns in fmcb_data that begin with 'fmcb' but are not scale items must be included here. Default is empty vector.}
}
\value{
If 'id' argument is used, returns a list with 2 dataframes: (1) bids_phenotype (contains input fmcb_data [values identical to input, underscores removed from fmcb items col names, if they existed] and FMCB scores) and (2) score_dat (contains FMCB subscale scores only). If 'id' argument is not used, returns a list with score_dat dataframe only.
}
\description{
This function scores the Feeding to Manage Child Behavior Questionnaire and provides subscale scores for the following behaviors: food to soothe and food as reward
}
\details{
For data to be scored correctly, the data must be prepared according to the following criteria: \cr
\itemize{
 \item{The data must include all individual questionnaire items}
 \item{The columns/variables must match the following naming convention: 'fmcb#' or 'fmcb_#' where # is the question number (1-10)}
 \item{All questionnaire responses must be a numeric value ranging from 0-4 (base_zero = TRUE) or 1-5 (base_zero = FALSE) where: }
 \itemize{
    \item{For base_zero = TRUE: 0 = Never; 1 = Rarely; 2 = Sometimes; 3 = Often; 4 = Always}
    \item{For base_zero = FALSE: 1 = Never; 2 = Rarely; 3 = Sometimes; 4 = Often; 5 = Always}
  }
 \item{Missing values must be coded as NA}
}
\cr
Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the fmcb Scale with IDs
fmcb_score_data <- score_fmcb(fmcb_data, base_zero = TRUE, id = 'ID')


}
\references{
Primary References for the Feeding to Manage Child Behavior Questionnaire and Scoring:
Savage JS, Ruggiero CF, Eagleton SG, Marini ME, Harris HA. The feeding to Manage Child Behavior Questionnaire: Development of a tool to measure' non-nutritive feeding practices in low income families with preschool-aged children. Appetite. 2022 Feb 1;169:105849. doi: 10.1016/j.appet.2021.105849. PMID: 34883138; PMCID: PMC8748389. (\href{https://pubmed.ncbi.nlm.nih.gov/34883138/}{PubMed})
}
