% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_hfi.R
\name{score_hfi}
\alias{score_hfi}
\title{score_hfi: Score data from the Fulkerson Home Food Inventory}
\usage{
score_hfi(hfi_data, base_zero = TRUE, id, session_id, extra_scale_cols = c())
}
\arguments{
\item{hfi_data}{a data.frame all items for the Fulkerson Home Food Inventory following the naming conventions described above}

\item{base_zero}{(logical) TRUE indicates the smallest value assigned to a choice is 0. FALSE indicates indicates the smallest value assigned to a choice is 1. Default = TRUE.}

\item{id}{(optional) name of participant ID column in input data. If included, the output dataset will be matched by id, if not included the output dataset will be in the order of the input data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}

\item{session_id}{(optional) name of session ID column in input data if there are multiple observations per participant. If included, the output dataset will be matched by id and session id, if not included the output dataset will be in the order of the input data. Required to get the phenotype dataset if have multiple sessions (raw data merged with scores.)}

\item{extra_scale_cols}{a vector of character strings that begin with 'bes' but are not scale items. Any columns in bes_data that begin with 'bes' but are not scale items must be included here. Default is empty vector.}
}
\value{
A dataset with subscale scores for the Fulkerson Home Food Inventory
}
\description{
This function scores the Fulkerson Home Food Inventory and provides subscale scores for: 1) 13 categories of food and 2) accessibility categories for kitchen and fridge (each sub-divided into unhealthy and healthy), and 3) an obesogenic score
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention: 'hfi_category_letter_type#' where 'letter' is the food item order and # type (where relevant)
3) All questions must have the numeric value for the choice with the base value being either 0 (base_zero = TRUE) or 1 (base_zero = FALSE)

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the hfi with IDs
hfi_score_data <- score_hfi(hfi_data, id = 'ID')


}
\references{
Fulkerson JA, Nelson MC, Lytle LA, Moe S, Heitzler C, Pasch KE. The validation of a home food inventory. International Journal of Behavioral Nutrition and Physical Activity, 2008, 5;55 (\href{https://pubmed.ncbi.nlm.nih.gov/18983668/}{PubMed})
}
