% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_hfias.R
\name{score_hfias}
\alias{score_hfias}
\title{score_hfias: Scored data from the Household Food Insecurity Access Scale}
\usage{
score_hfias(
  hfias_data,
  pna_value,
  base_zero = TRUE,
  id,
  session_id,
  extra_scale_cols = c()
)
}
\arguments{
\item{hfias_data}{a data.frame all items for the Pictorial Personality Traits Questionnaire for Children  following the naming conventions described above}

\item{pna_value}{value used when participant prefers not to answer/elects to skip}

\item{base_zero}{(logical) TRUE indicates the smallest value assigned to a choice is 0. FALSE indicates indicates the smallest value assigned to a choice is 1. Default = TRUE.}

\item{id}{(optional) name of participant ID column in input data. If included, the output dataset will be matched by id, if not included the output dataset will be in the order of the input data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}

\item{session_id}{(optional) name of session ID column in input data if there are multiple observations per participant. If included, the output dataset will be matched by id and session id, if not included the output dataset will be in the order of the input data. Required to get the phenotype dataset if have multiple sessions (raw data merged with scores.)}

\item{extra_scale_cols}{a vector of character strings that begin with 'hfias' but are not scale items. Any columns in sic_data that begin with 'hfias' but are not scale items must be included here. Default is empty vector.}
}
\value{
A dataset with subscale scores for the Pictorial Personality Traits Questionnaire for Children
}
\description{
This function scores the Household Food Insecurity Access Scale and provides continuous and categorical indicators of food security
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention 'hfias#' or 'hfias_#' for categorical questions and 'hfias#a' or 'hfias_#a' for frequency questions, where # is the question number (1-9)
3) Categorical questions 1-9 must have the numeric value for the choices: 0 - no (skips frequency question), 1 - yes
   Frequency Questions must have the numeric value for the choices: 0 - Rarely, 1 - Sometimes, 2 - Often (base_zero = TRUE) or 1 - Rarely, 2 - Sometimes, 3 - Often (base_zero = FALSE)

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the hfias with IDs
hfias_score_data <- score_hfias(hfias_data, id = 'ID')


\dontrun{
}

}
\references{
Primary Reference for the Questionnaire and Scoring:
Coates, Jennifer, Anne Swindale and Paula Bilinsky. 2007. Household Food Insecurity Access Scale (HFIAS) for Measurement of Household Food Access: Indicator Guide (v. 3). Washington, D.C.:FHI 360/FANTA.
}
