% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_hfssm.R
\name{score_hfssm}
\alias{score_hfssm}
\title{score_hfssm: Scored data from the U.S. Household Food Security Survey Module}
\usage{
score_hfssm(
  hfssm_data,
  pna_value,
  base_zero = FALSE,
  id,
  session_id,
  extra_scale_cols = c()
)
}
\arguments{
\item{hfssm_data}{a data.frame all items for the U.S. Household Food Security Survey Module following the naming conventions described above}

\item{pna_value}{value used when participant prefers not to answer/elects to skip}

\item{base_zero}{(logical) TRUE indicates the smallest value assigned to a choice is 0. FALSE indicates indicates the smallest value assigned to a choice is 1. Default = TRUE.}

\item{id}{(optional) name of participant ID column in input data. If included, the output dataset will be matched by id, if not included the output dataset will be in the order of the input data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}

\item{session_id}{(optional) name of session ID column in input data if there are multiple observations per participant. If included, the output dataset will be matched by id and session id, if not included the output dataset will be in the order of the input data. Required to get the phenotype dataset if have multiple sessions (raw data merged with scores.)}

\item{extra_scale_cols}{a vector of character strings that begin with 'hfssm' but are not scale items. Any columns in scpf_data that begin with 'hfssm' but are not scale items must be included here. Default is empty vector.}
}
\value{
A dataset with subscale scores for the U.S. Household Food Security Survey Module
}
\description{
This function scores the U.S. Household Food Security Survey Module and provides scores for: U.S. Household Food Security Scale, U.S. Adult Food Security Scale, 6-item Food Security Scale, U.S. Children’s Food Security Scale
}
\details{
To use this function, the data must be prepared according to the following criteria: \cr
\cr
1) The data must include all individual questionnaire items, excluding hh1, which is not used for scoring food security scales
\cr
2) The  columns/variables must match the following naming conventions: 'hfssm_hh#', 'hfssm_ad#', or 'hfssm_ch#', where # is the question number and hh/ad/ch refers to the questionnaire stage (household, adult, child)
\cr
3a) If base_zero = FALSE (default), questions must have the numeric value for the choices: \cr
\itemize{
 \item{Questions [hh2, hh3, hh4, ch1, ch2, ch3]: 1 - Often True, 2 - Sometimes True, 3 - Never True, 4 - DK or Refused}
 \item{Questions [ad1, ad2, ad3, ad4, ad5, ch4, ch5, ch6, ch7]: 1 - yes, 2 - no, 3 - DK}
 \item{Questions [ad1a, ad5a, ch5a]: 1 - almost every month, 2 - Some months but not every month, 3 - Only 1 or 2 months, 4 - DK}
}
3b) If base_zero = TRUE, questions must have the numeric value for the choices: \cr
\itemize{
 \item{Questions [hh2, hh3, hh4, ch1, ch2, ch3]: 0 - Often True, 1 - Sometimes True, 2 - Never True, 3 - DK or Refused}
 \item{Questions [ad1, ad2, ad3, ad4, ad5, ch4, ch5, ch6, ch7]: 0 - yes, 1 - no, 2 - DK}
 \item{Questions [ad1a, ad5a, ch5a]: 0 - almost every month, 1 - Some months but not every month, 2 - Only 1 or 2 months, 3 - DK}
}

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the hfssm with IDs
hfssm_score_data <- score_hfssm(hfssm_data, id = 'ID')


\dontrun{
}

}
\references{
U.S. HOUSEHOLD FOOD SECURITY SURVEY MODULE: THREE-STAGE DESIGN, WITH SCREENERS Economic Research Service, USDA September 2012 https://www.ers.usda.gov/media/8271/hh2012.pdf
}
