% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_nik_hfe.R
\name{score_nik_hfe}
\alias{score_nik_hfe}
\title{score_nik_hfe: Scored data from the Home Food Environment Survey from the Neighborhood Impact on Kids Study}
\usage{
score_nik_hfe(hfe_data, base_zero = TRUE, id, extra_scale_cols = c())
}
\arguments{
\item{hfe_data}{a data.frame all items for the Home Food Environment Survey from the Neighborhood Impact on Kids Study following the naming conventions described above}

\item{base_zero}{(logical) TRUE indicates the smallest value assigned to a choice is 0. FALSE indicates indicates the smallest value assigned to a choice is 1. Default = TRUE.}

\item{id}{(optional) name of participant ID column in input data. If included, the output dataset will be matched by id, if not included the output dataset will be in the order of the input data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}

\item{extra_scale_cols}{a vector of character strings that begin with 'hfe' but are not scale items. Any columns in hfe_data that begin with 'hfe' but are not scale items must be included here. Default is empty vector.}
}
\value{
A dataset with subscale scores for the Home Food Environment Survey from the Neighborhood Impact on Kids Study
}
\description{
This function scores the Home Food Environment Survey from the Neighborhood Impact on Kids Study and provides subscale scores for many different sections related to parent and child behaviors and neighborhood environment
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention: 'hfe#' or 'hfe_#' where # is the question number
3) All questions must have the numeric value for the choice: 0 - Never True, 1 - Rarely, 2 - Sometimes, 3 - Often, 4 - Always

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the hfe with IDs
hfe_score_data <- score_hfe(hfe_data, id = 'ID', base_zero = TRUE)

\dontrun{
}


}
\references{
B1. Couch SC, Glanz K, Zhou C, Sallis JF, Saelens BE. Home Food Environment in Relation to Children’s Diet Quality and Weight Status. Journal of the Academy of Nutrition and Dietetics. 2014;114(10):1569-1579.e1. doi:10.1016/j.jand.2014.05.015 (\href{https://pubmed.ncbi.nlm.nih.gov/25066057/}{PubMed})
}
