% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_pptq.R
\name{score_pptq}
\alias{score_pptq}
\title{score_pptq: Scored data from the Pictorial Personality Traits Questionnaire for Children}
\usage{
score_pptq(pptq_data, pptq_scale, score_base = TRUE, id)
}
\arguments{
\item{pptq_data}{a data.frame all items for the Pictorial Personality Traits Questionnaire for Children  following the naming conventions described above}

\item{pptq_scale}{3 or 5: indicates if data was collected with a 3-point or 5-point likert scale}

\item{score_base}{the smallest value assigned to a choice is 0 (i.e., range 0-3). Default = TRUE.}

\item{id}{(optional) name of participant ID column in bes_data. If included the output dataset will be matched by id, if not included the output dataset will be in the order of bes_data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}
}
\value{
A dataset with subscale scores for the Pictorial Personality Traits Questionnaire for Children
}
\description{
This function scores the Pictorial Personality Traits Questionnaire for Children and provides subscale scores for: Extraversion, Neuroticism, Openness, Conscientiousness, Agreeableness
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention: 'pptq#' or 'pptq_#' where # is the question number (1-15)
3) Questions 1-15 must have the numeric value for the choices: 1 - Definitely Yes (left side), 2 - A little bit (left side), 3 - It depends, 4 - A little bit (right side), 5 - Definitely yes (right side) if pptq_scale = 5 or
                                                               1 - (left side), 2 - It depends (middle), 3 - (right side) if pptq_scale = 3

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the pptq with IDs
pptq_score_data <- score_pptq(pptq_data, id = 'ID')


\dontrun{
}

}
\references{
Primary Reference for the Questionnaire and Scoring:
Maćkiewicz M, Cieciuch J. Pictorial Personality Traits Questionnaire for Children (PPTQ-C)-A New Measure of Children's Personality Traits. Front Psychol. 2016 Apr 14;7:498. doi: 10.3389/fpsyg.2016.00498. PMID: 27252661; PMCID: PMC4879772.
}
