% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_pss.R
\name{score_pss}
\alias{score_pss}
\title{score_pss: Scored data from the 10-item Perceived Stress Scale}
\usage{
score_pss(
  pss_data,
  pna_value,
  base_zero = TRUE,
  id,
  session_id,
  extra_scale_cols = c()
)
}
\arguments{
\item{pss_data}{a data.frame all items for the Perceived Stress Scale following the naming conventions described above}

\item{pna_value}{value used when participant prefers not to answer/elects to skip}

\item{base_zero}{(logical) TRUE indicates the smallest value assigned to a choice is 0. FALSE indicates indicates the smallest value assigned to a choice is 1. Default = TRUE.}

\item{id}{(optional) name of participant ID column in input data. If included, the output dataset will be matched by id, if not included the output dataset will be in the order of the input data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}

\item{session_id}{(optional) name of session ID column in input data if there are multiple observations per participant. If included, the output dataset will be matched by id and session id, if not included the output dataset will be in the order of the input data. Required to get the phenotype dataset if have multiple sessions (raw data merged with scores.)}

\item{extra_scale_cols}{a vector of character strings that begin with 'pss' but are not scale items. Any columns in pss_data that begin with 'pss' but are not scale items must be included here. Default is empty vector.}
}
\value{
A dataset with subscale scores for the Perceived Stress Scale
}
\description{
This function scores the 10-item Perceived Stress Scale and provides a total PSS score and subscale scores for: Perceived Helplessness and Perceived Self-efficacy
}
\details{
For data to be scored correctly, the data must be prepared according to the following criteria: \cr
\itemize{
 \item{The data must include all individual questionnaire items}
 \item{The columns/variables must match the following naming convention: 'pss#' or 'pss_#' where # is the question number (1-10)}
 \item{All questionnaire responses must be a numeric value ranging from 0-4 (base_zero = TRUE) or 1-5 (base_zero = FALSE) where: }
 \itemize{
    \item{For base_zero = TRUE: 0 = Never; 1 = Almost Never; 2 = Sometimes; 3 = Fairly Often; 4 = Very Often}
    \item{For base_zero = FALSE: 1 = Never; 2 = Almost Never; 3 = Sometimes; 4 = Fairly Often; 5 = Very Often}
  }
 \item{Missing values must be coded as NA}
 \item{Values must not be reversed scored. This script will apply reverse scoring so all levels must be true to the scale described above}
}
\cr
Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the pss with IDs
pss_score_data <- score_pss(pss_data, id = 'ID')


\dontrun{
}

}
\references{
Cohen, S., Kamarck, T., and Mermelstein, R.(1983). A global measure of perceived stress. Journal of Health and Social Behavior, 24, 386-396. PMID: 6668417.
Cohen, S. and Williamson, G. Perceived Stress in a Probability Sample of the United States. Spacapan, S. and Oskamp, S. (Eds.) The Social Psychology of Health. Newbury Park, CA: Sage, 1988.


Two-factor subscale scoring comes from:
Taylor JM. Psychometric analysis of the Ten-Item Perceived Stress Scale. Psychol Assess. 2015 Mar;27(1):90-101. doi: 10.1037/a0038100. Epub 2014 Oct 27. PMID: 25346996.
}
