% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_pwlb.R
\name{score_pwlb}
\alias{score_pwlb}
\title{score_pwlb: Scored data from the Parent Weight-Loss Behavior Questionnaire}
\usage{
score_pwlb(pwlb_data, base_zero = TRUE, id, session_id, extra_scale_cols = c())
}
\arguments{
\item{pwlb_data}{a data.frame all items for the Parent Weight-Loss Behavior Questionnaire following the naming conventions described above}

\item{base_zero}{(logical) TRUE indicates the smallest value assigned to a choice is 0. FALSE indicates indicates the smallest value assigned to a choice is 1. Default = TRUE.}

\item{id}{(optional) name of participant ID column in input data. If included, the output dataset will be matched by id, if not included the output dataset will be in the order of the input data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}

\item{session_id}{(optional) name of session ID column in input data if there are multiple observations per participant. If included, the output dataset will be matched by id and session id, if not included the output dataset will be in the order of the input data. Required to get the phenotype dataset if have multiple sessions (raw data merged with scores.)}

\item{extra_scale_cols}{a vector of character strings that begin with 'pwlb' but are not scale items. Any columns in pwlb_data that begin with 'pwlb' but are not scale items must be included here. Default is empty vector.}
}
\value{
A dataset with subscale scores for the Parent Weight-Loss Behavior Questionnaire
}
\description{
This function scores the Parent Weight-Loss Behavior Questionnaire and provides subscale scores for the following behaviors: Healthy Weight Control and Unhealthy Weight Control.
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention: 'pwlb#' or 'pwlb_#' where # is the question number (1-29; only 1-24 are needed to compute scale as 25-29 are contextual questions)
3) Questions 1-24 must have the numeric value for the choices: 1 - Never, 2 - Rarely, 3 - Sometimes, 4 - Often, 5 - Always. Questions 25-29 are scored differently and are follow-up questions.

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the pwlb with IDs
pwlb_score_data <- score_pwlb(pwlb_data, id = 'ID')


\dontrun{
}



}
\references{
Primary References for the Parent Weight-Loss Behavior Questionnaire and Scoring:
Savage JS, Birch LL. Patterns of weight control strategies predict differences in women’s 4 y weight gain. Obesity (Silver Spring). 2010;18(3):513-520. doi:10.1038/oby.2009.265 (\href{https://pubmed.ncbi.nlm.nih.gov/19696759/}{PubMed})

Measure Adapted from:
French SA, Perry CL, Leon GR, Fulkerson JA. Dieting behaviors and weight change history in female adolescents. Health Psychology. 1995;14(6):548-555. doi:http://dx.doi.org/10.1037/0278-6133.14.6.548 (\href{https://pubmed.ncbi.nlm.nih.gov/8565929/}{PubMed})
}
