% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_sdor.R
\name{score_sdor}
\alias{score_sdor}
\title{score_sdor: Scored data from the Scatter Division of Responsibility in Feeding Scale}
\usage{
score_sdor(
  sdor_data,
  pna_value,
  base_zero = TRUE,
  id,
  session_id,
  extra_scale_cols = c()
)
}
\arguments{
\item{sdor_data}{a data.frame all items for the Scatter Division of Responsibility in Feeding Scale following the naming conventions described above}

\item{pna_value}{value used when participant prefers not to answer/elects to skip}

\item{base_zero}{(logical) TRUE indicates the smallest value assigned to a choice is 0. FALSE indicates indicates the smallest value assigned to a choice is 1. Default = TRUE.}

\item{id}{(optional) name of participant ID column in input data. If included, the output dataset will be matched by id, if not included the output dataset will be in the order of the input data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}

\item{session_id}{(optional) name of session ID column in input data if there are multiple observations per participant. If included, the output dataset will be matched by id and session id, if not included the output dataset will be in the order of the input data. Required to get the phenotype dataset if have multiple sessions (raw data merged with scores.)}

\item{extra_scale_cols}{a vector of character strings that begin with 'sdor' but are not scale items. Any columns in sdor_data that begin with 'sdor' but are not scale items must be included here. Default is empty vector.}
}
\value{
A dataset with subscale scores for the Scatter Division of Responsibility in Feeding Scale
}
\description{
This function scores the Scatter Division of Responsibility in Feeding Scale
}
\details{
For data to be scored correctly, the data must be prepared according to the following criteria: \cr
\itemize{
 \item{The data must include all individual questionnaire items except for question 13:}
 \item{The columns/variables must match the following naming convention: 'sdor#' or 'sdor_#' where # is the question number (1-15).}
 \item{All questionnaire responses must be a numeric value ranging from 0-4 (base_zero = TRUE) or 1-5 (base_zero = FALSE) where: }
 \itemize{
   \item{For base_zero = TRUE: 0 = Never; 1 = Rarely; 2 = Sometimes; 3 = Often; 4 = Always}
   \item{For base_zero = TRUE: 1 = Never; 2 = Rarely; 3 = Sometimes; 4 = Often; 5 = Always}
  }
 \item{Missing values must be coded as NA}
}
\cr
Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the sdor with IDs, with scale from 0-1, prefer not to answer indicated with 2
sdor_score_data <- score_sdor(sdor_data, base_zero = TRUE, id = 'ID', pna_value = 2)


\dontrun{
}


}
\references{
Lohse B, Satter E, Arnold K. Development of a tool to assess adherence to a model of the division of responsibility in feeding young children: using response mapping to capacitate validation measures. Child Obes. 2014 Apr;10(2):153-68. doi: 10.1089/chi.2013.0085. PMID: 24716583. (\href{https://pubmed.ncbi.nlm.nih.gov/24716583/}{PubMed})
}
