% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_sic.R
\name{score_sic}
\alias{score_sic}
\title{score_sic: INCOMPLETE FUNCTION Scored data from the Stress in Children Questionnaire}
\usage{
score_sic(sic_data, score_base = TRUE, id, extra_scale_cols = c())
}
\arguments{
\item{sic_data}{a data.frame all items for the Child Behavior Questionnaire following the naming conventions described above}

\item{score_base}{the smallest value assigned to a choice is 0 (i.e., range 0-3). Default = TRUE.}

\item{id}{(optional) name of participant ID column in bes_data. If included the output dataset will be matched by id, if not included the output dataset will be in the order of bes_data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}

\item{extra_scale_cols}{a vector of character strings that begin with 'sic' but are not scale items. Any columns in sic_data that begin with 'sic' but are not scale items must be included here. Default is empty vector.}
}
\value{
A dataset with subscale scores for the Stress in Chidlren Questionnaire
}
\description{
This function scores the Stress in Children Questionnaire and provides subscale scores for the following behaviors: Well Being, Distress, Social Support, and the Global Mean Score.
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention: 'sic#' or 'sic_#' where # is the question number (1-21)
3) All questions must have the numeric value for the choice: 1 - Never, 2 - Sometimes, 3 - Often, 4 - Very Often

Note, as long as variable names match those listed, the dataset can include other variables. Up to 2 missing responses are allowed
}
\examples{

# scoring for the sic with IDs
sic_score_data <- score_sic(sic_data, id = 'ID')

\dontrun{
}


}
\references{
1. Osika W, Friberg P, Wahrborg P. A new short self-rating questionnaire to assess stress in children. Int J Behav Med. 2007;14(2):108-117. doi:10.1007/BF03004176 (\href{https://pubmed.ncbi.nlm.nih.gov/17926439/}{PubMed})
}
