% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_sleeplog.R
\name{score_sleeplog}
\alias{score_sleeplog}
\title{score_sleeplog: Score data from the week-long sleep log}
\usage{
score_sleeplog(sleep_data, id, session_id, summer_start, summer_end)
}
\arguments{
\item{sleep_data}{a data.frame all items for the the week-long sleep log following the naming conventions described above}

\item{id}{(optional) name of participant ID column in input data. If included, the output dataset will be matched by id, if not included the output dataset will be in the order of the input data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}

\item{session_id}{(optional) name of session ID column in input data if there are multiple observations per participant. If included, the output dataset will be matched by id and session id, if not included the output dataset will be in the order of the input data. Required to get the phenotype dataset if have multiple sessions (raw data merged with scores.)}

\item{summer_start}{a date for the start of summer holiday (check area school districts during data collection). Must be formatted as character string 'YYYY-MM-DD'.}

\item{summer_end}{a date for the end of summer holiday (check area school districts during data collection). Must be formatted as character string 'YYYY-MM-DD'.}
}
\value{
A dataset with total score for the the week-long sleep log
}
\description{
This function scores the week-long sleep log used for study BRAKE during the week children wore an actigraphy watch
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The columns/variables must match the following naming convention: 'var_day' where var is the variable measured (e.g., bedtime, date, etc) and day is the abbreviation for day of the week (mon, tu, wed, th, fri, sat, sun)

Note, as long as variable names match those listed in this script, the dataset can include other variables
}
\examples{

# scoring for the sleep log with IDs
sleep_score_data <- score_sleeplog(sleep_data, id = 'ID', summer_start = '2023-06-06', summer_end = '2023-08-23')

\dontrun{
} 


}
