% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_socialdq.R
\name{score_socialdq}
\alias{score_socialdq}
\title{score_socialdq: Score data from the Social Desirability Questionnaire}
\usage{
score_socialdq(
  sdq_data,
  pna_value,
  id,
  session_id,
  base_zero = TRUE,
  extra_scale_cols = c()
)
}
\arguments{
\item{sdq_data}{a data.frame all items for the Social Desirability Questionnaire following the naming conventions described above}

\item{pna_value}{(integer) integer used for items where child 'prefers not to answer'}

\item{id}{(optional) name of participant ID column in input data. If included, the output dataset will be matched by id, if not included the output dataset will be in the order of the input data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}

\item{session_id}{(optional) name of session ID column in input data if there are multiple observations per participant. If included, the output dataset will be matched by id and session id, if not included the output dataset will be in the order of the input data. Required to get the phenotype dataset if have multiple sessions (raw data merged with scores.)}

\item{base_zero}{(logical) TRUE indicates the smallest value assigned to a choice is 0. FALSE indicates indicates the smallest value assigned to a choice is 1. Default = TRUE.}

\item{extra_scale_cols}{vector of character strings that begin with 'sdq' but are not scale items. Any columns in sdq_data that begin with 'sdq' but are not scale items must be included here. Default is empty vector.}
}
\value{
A dataset with a score for the Social Desirability Questionnaire
}
\description{
This function scores the Social Desirability Questionnaire
}
\details{
To use this function, the data must be prepared according to the following criteria: \cr
\itemize{
 \item{The data must include all individual questionnaire items}
 \item{The  columns/variables must match the following naming convention: 'sdq#' or 'sdq_# where # is the question number (1-28)}
 \item{All questions must have the numeric value for the choice starting with the value 0 (base_zero = TRUE) or 1 (base_zero = FALSE)}
}
Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the sdq with IDs
sdq_score_data <- score_socialdq(sdq_data, parID = 'ID')

\dontrun{
}


}
\references{
Ford LH. A social desirability questionnaire for young children. Journal of Consulting and Clinical Psychology 1970; 35(2):195-204.
}
