% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_cebq.R
\name{score_cebq}
\alias{score_cebq}
\title{score_cebq: Scored data from the Children's Eating Behavior Questionnaire}
\usage{
score_cebq(
  cebq_data,
  pna_value,
  base_zero = TRUE,
  id,
  session_id,
  extra_scale_cols = c()
)
}
\arguments{
\item{cebq_data}{a data.frame all items for the Children's Eating Behavior Questionnaire following the naming conventions described in Details}

\item{pna_value}{value used when participant prefers not to answer/elects to skip}

\item{base_zero}{(logical) TRUE indicates the smallest value assigned to a choice is 0. FALSE indicates indicates the smallest value assigned to a choice is 1. Default = TRUE.}

\item{id}{(optional) name of participant ID column in input data. If included, the output dataset will be matched by id, if not included the output dataset will be in the order of the input data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}

\item{session_id}{(optional) name of session ID column in input data if there are multiple observations per participant. If included, the output dataset will be matched by id and session id, if not included the output dataset will be in the order of the input data. Required to get the phenotype dataset if have multiple sessions (raw data merged with scores.)}

\item{extra_scale_cols}{a vector of character strings that begin with 'cebq' but are not scale items. Any columns in cebq_data that begin with 'cebq' but are not scale items must be included here. Default is empty vector.}
}
\value{
If 'id' argument is used, returns a list with 2 dataframes: (1) bids_phenotype (contains input cebq_data [values identical to input, underscores removed from cebq items col names, if they existed] and cebq subscale scores) and (2) score_dat (contains cebq subscale scores only). If 'id' argument is not used, returns a list with score_dat dataframe only.
}
\description{
This function scores the Children's Eating Behavior Questionnaire and provides subscale scores for the following behaviors: Food Responsiveness, Emotional Overeating, Enjoyment of Food, Desire to Drink, Satiety Responsiveness, Slowness in Eating, Emotional Undereating, and Food Fussiness.
}
\details{
Recent analyses by Manzano et al., (2021) indicate a better-fitting 3 factor solution for school-age children, therefore, the following subscales are also reported: reward-based eating, picky eating, and emotional eating

For data to be scored correctly, the data must be prepared according to the following criteria: \cr
\itemize{
 \item{The data must include all individual questionnaire items}
 \item{The columns/variables must match the following naming convention: 'cebq#' or 'cebq_#' where # is the question number (1-35)}
 \item{All questionnaire responses must be a numeric value ranging from 0-4 (base_zero = TRUE) or 1-5 (base_zero = FALSE) where: }
 \itemize{
    \item{For base_zero = TRUE: 0 = Never; 1 = Rarely; 2 = Sometimes; 3 = Often; 4 = Always}
    \item{For base_zero = FALSE: 1 = Never; 2 = Rarely; 3 = Sometimes; 4 = Often; 5 = Always}
  }
 \item{Missing values must be coded as NA}
 \item{Values must not be reversed scored. This script will apply reverse scoring so all levels must be true to the scale described above}
}
\cr
Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the CEBQ with IDs, when values range from 0-4
data(cebq_base0) # load example data included with dataprepr
cebq_score_data <- score_cebq(cebq_data = cebq_base0, base_zero = TRUE, id = 'ID')

# scoring for the CEBQ with IDs, when values range from 1-5
data(cebq_base1) # load example data included with dataprepr
cebq_score_data <- score_cebq(cebq_data = cebq_base1, base_zero = FALSE, id = 'ID')

}
\references{
Primary References for the Children's Eating Behavior Questionnaire and Scoring:
Wardle, J., Guthrie, C. A., Sanderson, S., & Rapoport, L. (2001). Development of the children’s eating behaviour questionnaire. Journal of Child Psychology and Psychiatry, 42, 963–970. https://doi.org/10.1017/S0021963001007727 (\href{https://pubmed.ncbi.nlm.nih.gov/11693591/}{PubMed})

Alternative 3-factor scoring:
Manzano MA, Strong DR, Kang Sim DE, Rhee KE, Boutelle KN. Psychometric properties of the Child Eating Behavior Questionnaire (CEBQ) in school age children with overweight and obesity: A proposed three‐factor structure. Pediatric Obesity. 2021;16(10):e12795. doi:10.1111/ijpo.12795 (\href{https://pubmed.ncbi.nlm.nih.gov/33945226/}{PubMed})'
}
