% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_cfpq.R
\name{score_cfpq}
\alias{score_cfpq}
\title{score_cfpq: Scored data from the Comprehensive Feeding Practices Questionnaire}
\usage{
score_cfpq(cfpq_data, base_zero = TRUE, id, extra_scale_cols = c())
}
\arguments{
\item{cfpq_data}{a data.frame all items for the Comprehensive Feeding Practices Questionnaire following the naming conventions described above}

\item{base_zero}{(logical) TRUE indicates the smallest value assigned to a choice is 0. FALSE indicates indicates the smallest value assigned to a choice is 1. Default = TRUE.}

\item{id}{(optional) name of participant ID column in input data. If included, the output dataset will be matched by id, if not included the output dataset will be in the order of the input data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}

\item{extra_scale_cols}{a vector of character strings that begin with 'cfpq' but are not scale items. Any columns in cfpq_data that begin with 'cfpq' but are not scale items must be included here. Default is empty vector.}
}
\value{
A dataset with subscale scores for the Comprehensive Feeding Practices Questionnaire
}
\description{
This function scores the Comprehensive Feeding Practices Questionnaire and provides subscale scores for: child control, emotional regulation, Encourage balance and variety, environment, food as reward, involvement, modeling, monitoring, pressure, restriction for health, restriction for weight control, and teaching about nutrition
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention: 'cfpq#' or 'cfpq_#' where # is the question number (1-49)
3) Questions 1-13 must have the numeric value for the choices: 1 - never, 2 - rarely, 3 - sometimes, 4 - mostly, 5 - always
   Questions 14-49 must have the numeric value for the choices: 1 - disagree, 2 - slightly disagree, 3 - neutral, 4 - slightly agree, 5 - agree.

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the cfpq with IDs
cfpq_score_data <- score_cfpq(cfpq_data, id = 'ID')


\dontrun{
}

}
\references{
Primary Reference for the Questionnaire and Scoring:
Musher-Eizenman, D., Holub. S., Comprehensive Feeding Practices Questionnaire: Validation of a New Measure of Parental Feeding Practices, Journal of Pediatric Psychology, 32, 8, September 2007, Pages 960–972, https://doi.org/10.1093/jpepsy/jsm037
}
