% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_cfq.R
\name{score_cfq}
\alias{score_cfq}
\title{score_cfq: Scored data from the Child Feeding Questionnaire}
\usage{
score_cfq(
  cfq_data,
  pcw_na_value,
  base_zero = TRUE,
  restriction_split = FALSE,
  id,
  session_id,
  extra_scale_cols = c()
)
}
\arguments{
\item{cfq_data}{a data.frame all items for the Child Feeding Questionnaire following the naming conventions described above}

\item{pcw_na_value}{(integer) integer used for items 8-13 (Perceived Child Weight items) to indicate "not applicable" response.}

\item{base_zero}{(logical) TRUE indicates the smallest value assigned to a choice is 0. FALSE indicates indicates the smallest value assigned to a choice is 1. Default = TRUE.}

\item{restriction_split}{logical indicating if the Restriction subscale should be split to remove food as reward items. Default = FALSE. The standard Restriction subscale will always be available. If restriction_split = TRUE, then two additional scales will be computed: 1) cfq_rest_noreward: questions 17-20, 23-24 and 1) cfq_foodreward: questions 21-22}

\item{id}{(optional) name of participant ID column in input data. If included, the output dataset will be matched by id, if not included the output dataset will be in the order of the input data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}

\item{session_id}{(optional) name of session ID column in input data if there are multiple observations per participant. If included, the output dataset will be matched by id and session id, if not included the output dataset will be in the order of the input data. Required to get the phenotype dataset if have multiple sessions (raw data merged with scores.)}

\item{extra_scale_cols}{a vector of character strings that begin with 'cfq' but are not scale items. Any columns in cfq_data that begin with 'cfq' but are not scale items must be included here. Default is empty vector.}
}
\value{
A dataset with subscale scores for the Child Feeding Questionnaire
}
\description{
This function scores the Child Feeding Questionnaire and provides subscale scores for the following behaviors: Perceived Responsibility, Perceived Child Weight, Perceived Parent Weight, Child Weight Concerns, Restriction, Pressure to Eat, and Monitoring
}
\details{
For all subscales except Perceived Child Weight, scores will only be computed if there are no missing values in subscale items. For Perceived Child Weight, subscale scores will be calculated ignoring missing items and pcw_na_value responses (see Arguments) 

For data to be scored correctly, the data must be prepared according to the following criteria: \cr
\itemize{
 \item{The data must include all individual questionnaire items except for question 13:}
 \itemize{
   \item{Question 13 can be excluded from data due to age range. If excluded, Perceived Child Weight will be calculated using items 8-12}
  }
 \item{The columns/variables must match the following naming convention: 'cfq#' or 'cfq_#' where # is the question number (1-31).}
 \item{All questionnaire responses must be a numeric value ranging from 0-4 (base_zero = TRUE) or 1-5 (base_zero = FALSE) where: }
 \itemize{
   \item{For base_zero = TRUE:}
   \itemize{
     \item{Items 1-3: 0 = Never; 1 = Seldom; 2 = Half of time; 3 = Most of time; 4 = Always}
     \item{Items 4-13: 0 = Markedly Underweight; 1 = Underweight; 2 = Average; 3 = Overweight; 4 = Markedly Overweight}
     \item{Items 14-16: 0 = Unconcerned; 1 = Slightly unconcerned; 2 = Neutral; 3 = Slightly concerned; 4 = Very Concerned}
     \item{Items 17-28: 0 = Disagree; 1 = Slightly disagree; 2 = Neutral; 3 = Slightly agree; 4 = Agree}
     \item{Items 29-31: 0 = Never; 1 = Rarely; 2 = Sometimes; 3 = Mostly; 4 = Always}
   }
   \item{For base_zero = TRUE:}
   \itemize{
     \item{Items 1-3: 1 = Never; 2 = Seldom; 3 = Half of time; 4 = Most of time; 5 = Always}
     \item{Items 4-13: 1 = Markedly Underweight; 2 = Underweight; 3 = Average; 4 = Overweight; 5 = Markedly Overweight}
     \item{Items 14-16: 1 = Unconcerned; 2 = Slightly unconcerned; 3 = Neutral; 4 = Slightly concerned; 5 = Very Concerned}
     \item{Items 17-28: 1 = Disagree; 2 = Slightly disagree; 3 = Neutral; 4 = Slightly agree; 5 = Agree}
     \item{Items 29-31: 1 = Never; 2 = Rarely; 3 = Sometimes; 4 = Mostly; 5 = Always}
   }
  }
 \item{Missing values must be coded as NA}
 \item{Items 8-13 (Perceived Child Weight subscale) can include responses reflecting "not applicable", as these items ask about certain age ranges. The value encoding "not applicable" must be specified with the pcw_na_value parameter}
}
\cr
Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the cfq with IDs, with scale from 0-4, and items 8-13 have the value 99 indicating "not applicable"
cfq_score_data <- score_cfq(cfq_data, base_zero = TRUE, id = 'ID', pcw_na_value = 99)

# scoring for the cfq with extra Restriction subscales, with scale from 1-5
cfq_score_data <- score_cfq(cfq_data, base_zero = FALSE, restriction_split = TRUE, id = 'ID')

\dontrun{
}


}
\references{
Birch, L. L., Fisher, J. O., Grimm-Thomas, K., Markey, C. N., Sawyer, R., & Johnson, S. L. (2001). Confirmatory factor analysis of the Child Feeng Questionnaire: A measure of parental attitudes, beliefs and practices about child feeding and obesity proneness. Appetite, 36(3), 201–210. https://doi.org/10.1006/appe.2001.0398 (\href{https://pubmed.ncbi.nlm.nih.gov/11358344/}{PubMed})
}
