% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_chaos.R
\name{score_chaos}
\alias{score_chaos}
\title{score_chaos: INCOMPLETE FUNCTION -- Scored data from the 15-item Confusion, Hubbub, and Order Scale}
\usage{
score_chaos(chaos_data, base_zero = TRUE, id, extra_scale_cols = c())
}
\arguments{
\item{chaos_data}{a data.frame all items for the Confusion, Hubbub, and Order Scale following the naming conventions described above}

\item{base_zero}{the smallest value assigned to a choice is 0 (i.e., range 0-3). Default = TRUE.}

\item{id}{(optional) name of participant ID column in bes_data. If included the output dataset will be matched by id, if not included the output dataset will be in the order of bes_data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}

\item{extra_scale_cols}{a vector of character strings that begin with 'chaos' but are not scale items. Any columns in chaos_data that begin with 'chaos' but are not scale items must be included here. Default is empty vector.}
}
\value{
A dataset with subscale scores for the Confusion, Hubbub, and Order Scale
}
\description{
This function scores the Confusion, Hubbub, and Order Scale and provides XXX
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention: 'chaos#' or 'chaos_#' where # is the question number (1-15).
3) Questions must have the numeric value for the choices: 1 - Very much like your own home, 2 - Somewhat like your own home, 3 - A little bit like your own home, 4 - Not at all like your own home

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the chaos with IDs
chaos_score_data <- score_chaos(chaos_data, id = 'ID')


\dontrun{
}



}
\references{
Matheny, A. P., Jr., Wachs, T. D., Ludwig, J. L., Phillips, K. (1995). Bringing order out of chaos: Psychometric characteristics of the confusion, hubbub, and order scale. Journal of Applied Developmental Psychology, 16(3), 429–444. https://doi.org/10.1016/0193-3973(95)90028-4
}
