% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_debq.R
\name{score_debq}
\alias{score_debq}
\title{score_debq: INCOMPLETE FUNCTION. Scored data from the Dutch Eating Behavior Questionnaire}
\usage{
score_debq(debq_data, score_base = TRUE, id)
}
\arguments{
\item{debq_data}{a data.frame all items for the Dutch Eating Behavior Questionnaire following the naming conventions described above}

\item{score_base}{the smallest value assigned to a choice is 0 (i.e., range 0-3). Default = TRUE.}

\item{id}{(optional) name of participant ID column in bes_data. If included the output dataset will be matched by id, if not included the output dataset will be in the order of bes_data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}
}
\value{
A dataset with subscale scores for the Dutch Eating Behavior Questionnaire
}
\description{
This function scores the Dutch Eating Behavior Questionnaire and provides subscale scores for the following behaviors:
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention: 'debq#' where # is the question number (1-24).
3) Questions must have the numeric value for the choices: 1=Never, 2=Rarely/Seldom, 3=Sometimes, 4=Often, 5=Very Often if score_base = FALSE or 0=Never, 1=Rarely/Seldom, 2=Sometimes, 3=Often, 4=Very Often if score_base = TRUE

Note, as long as variable names match those listed, the dataset can include other variables

Unlike the original Dutch version of the DEBQ (van Strien et al., 1986), the English version (Wardle, 1987) does not require reverse scoring of item 21. This function scores based on Wardle, 1987
}
\examples{

# scoring for the debq with IDs
debq_score_data <- score_debq(debq_data, id = 'ID')


\dontrun{
}


}
\references{
Wardle J. Eating style: a validation study of the Dutch Eating Behaviour Questionnaire in normal subjects and women with eating disorders. J Psychosom Res. 1987;31(2):161-9. doi: 10.1016/0022-3999(87)90072-9. PMID: 3473234.
}
