% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_debq.R
\name{score_debq}
\alias{score_debq}
\title{score_debq: Scored data from the Dutch Eating Behavior Questionnaire}
\usage{
score_debq(
  debq_data,
  base_zero = TRUE,
  id,
  session_id,
  na_handling = FALSE,
  extra_scale_cols = c()
)
}
\arguments{
\item{debq_data}{a data.frame all items for the Dutch Eating Behavior Questionnaire following the naming conventions described above}

\item{base_zero}{(logical) TRUE indicates the smallest value assigned to a choice is 0. FALSE indicates indicates the smallest value assigned to a choice is 1. Default = TRUE.}

\item{id}{(optional) name of participant ID column in input data. If included, the output dataset will be matched by id, if not included the output dataset will be in the order of the input data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}

\item{session_id}{(optional) name of session ID column in input data if there are multiple observations per participant. If included, the output dataset will be matched by id and session id, if not included the output dataset will be in the order of the input data. Required to get the phenotype dataset if have multiple sessions (raw data merged with scores.)}

\item{na_handling}{logical - remove NAs when computing subscales: TRUE will remove and comput means; FALSE will not remove NAs and subscale will not be computed if NA is present. Default is FALSE.}

\item{extra_scale_cols}{a vector of character strings that begin with 'debq' but are not scale items. Any columns in debq_data that begin with 'debq' but are not scale items must be included here. Default is empty vector.}
}
\value{
A dataset with subscale scores for the Dutch Eating Behavior Questionnaire
}
\description{
This function scores the Dutch Eating Behavior Questionnaire and provides subscale scores for: Emotional Eating, External Eating, Restrained Eating
}
\details{
For data to be scored correctly, the data must be prepared according to the following criteria: \cr
\itemize{
 \item{The data must include all individual questionnaire items}
 \item{The columns/variables must match the following naming convention: 'debq#' or 'debq_#' where # is the question number (1-24).}
 \item{All questionnaire responses must be a numeric value ranging from 0-4 (base_zero = TRUE) or 1-5 (base_zero = FALSE) where: }
 \itemize{
    \item{For base_zero = TRUE: 0 = Never; 1 = Rarely/Seldom; 2 = Sometimes; 3 = Often; 4 = Very Often}
    \item{For base_zero = FALSE: 1 = Never; 2 = Rarely/Seldom; 3 = Sometimes; 4 = Often; 5 = Very Often}
  }
 \item{Missing values must be coded as NA}
}
Note, as long as variable names match those listed, the dataset can include other variables

Unlike the original Dutch version of the DEBQ (van Strien et al., 1986), the English version (Wardle, 1987) does not require reverse scoring of item 21. This function scores based on Wardle, 1987
}
\examples{

# scoring for the debq with IDs
debq_score_data <- score_debq(debq_data, id = 'ID')


\dontrun{
}


}
\references{
Wardle J. Eating style: a validation study of the Dutch Eating Behaviour Questionnaire in normal subjects and women with eating disorders. J Psychosom Res. 1987;31(2):161-9. doi: 10.1016/0022-3999(87)90072-9. PMID: 3473234.
}
