% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_hfi.R
\name{score_hfi}
\alias{score_hfi}
\title{score_hfi: Score data from the Fulkerson Home Food Inventory}
\usage{
score_hfi(hfi_data, score_base = TRUE, id)
}
\arguments{
\item{hfi_data}{a data.frame all items for the Fulkerson Home Food Inventory following the naming conventions described above}

\item{score_base}{the smallest value assigned to a choice is 0 (i.e., range 0-3). Default = TRUE.}

\item{id}{(optional) name of participant ID column in bes_data. If included the output dataset will be matched by id, if not included the output dataset will be in the order of bes_data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}
}
\value{
A dataset with subscale scores for the ulkerson Home Food Inventory
}
\description{
This function scores the Fulkerson Home Food Inventory and provides subscale scores for: 1) 13 categories of food and 2) accessibility categories for kitchen and fridge (each sub-divided into unhealthy and healthy), and 3) an obesogenic score
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention: 'hfi_category_letter_type#' where 'letter' is the food item order and # type (where relevant)
3) All questions must have the numeric value for the choice with the base value being either 0 (score_base = TRUE) or 1 (score_base = FALSE)

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the hfi with IDs
hfi_score_data <- score_hfi(hfi_data, id = 'ID')


}
\references{
Fulkerson JA, Nelson MC, Lytle LA, Moe S, Heitzler C, Pasch KE. The validation of a home food inventory. International Journal of Behavioral Nutrition and Physical Activity, 2008, 5;55 (\href{https://pubmed.ncbi.nlm.nih.gov/18983668/}{PubMed})
}
