% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_hfssm.R
\name{score_hfssm}
\alias{score_hfssm}
\title{score_hfssm: Scored data from the U.S. Household Food Security Survey Module}
\usage{
score_hfssm(hfssm_data, base_zero = TRUE, id)
}
\arguments{
\item{hfssm_data}{a data.frame all items for the U.S. Household Food Security Survey Module following the naming conventions described above}

\item{base_zero}{(logical) TRUE indicates the smallest value assigned to a choice is 0. FALSE indicates indicates the smallest value assigned to a choice is 1. Default = TRUE.}

\item{id}{(optional) name of participant ID column in input data. If included, the output dataset will be matched by id, if not included the output dataset will be in the order of the input data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}
}
\value{
A dataset with subscale scores for the U.S. Household Food Security Survey Module
}
\description{
This function scores the U.S. Household Food Security Survey Module and provides scores for: U.S. Household Food Security Scale, U.S. Adult Food Security Scale, 6-item Food Security Scale, U.S. Children’s Food Security Scale
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items, excluding hh1, which is not used for scoring food security scales
2) The  columns/variables must match the following naming conventions: 'hfssm_hh#', 'hfssm_ad#', or 'hfssm_ch#', where # is the question number and hh/ad/ch refers to the questionnaire stage (household, adult, child)
3) Questions [hh2, hh3, hh4, ch1, ch2, ch3] must have the numeric value for the choices: 1 - Often True, 2 - Sometimes True, 3 - Never True, 4 - DK or Refused
   Questions [ad1, ad2, ad3, ad4, ad5, ch4, ch5, ch6, ch7] must have the numeric value for the choices: 1 - yes, 2 - no, 3 - DK
   Questions [ad1a, ad5a, ch5a] must have the numeric value for the choices: 1 - almost every month, 2 - Some months but not every month, 3 - Only 1 or 2 months, 4 - DK


Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the hfssm with IDs
hfssm_score_data <- score_hfssm(hfssm_data, id = 'ID')


\dontrun{
}

}
\references{
U.S. HOUSEHOLD FOOD SECURITY SURVEY MODULE: THREE-STAGE DESIGN, WITH SCREENERS Economic Research Service, USDA September 2012 https://www.ers.usda.gov/media/8271/hh2012.pdf
}
