% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_lbc.R
\name{score_lbc}
\alias{score_lbc}
\title{score_lbc: Scored data from the Lifestyle Behavior Checklist}
\usage{
score_lbc(lbc_data, score_base = TRUE, id)
}
\arguments{
\item{lbc_data}{a data.frame all items for the Lifestyle Behavior Checklist following the naming conventions described above}

\item{score_base}{the smallest value assigned to a choice is 0 (i.e., range 0-3). Default = TRUE.}

\item{id}{(optional) name of participant ID column in bes_data. If included the output dataset will be matched by id, if not included the output dataset will be in the order of bes_data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}
}
\value{
A dataset with subscale scores for the Lifestyle Behavior Checklist
}
\description{
This function scores the Lifestyle Behavior Checklist and provides subscale scores for the following behaviors: Food-Related Misbehavior, Overeating, Emotions Related to Overweight, and Physical Activity.
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention: 'lbc#' or 'lbc_#' where # is the question number (1-25)
3) All questions must have the numeric value for the choice with the base value being either 0 (score_base = TRUE) or 1 (score_base = FALSE)

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the lbc with IDs
lbc_score_data <- score_lbc(lbc_data, id = 'ID')


}
\references{
West F, Sanders MR. The Lifestyle Behaviour Checklist: A measure of weight-related problem behaviour in obese children. International Journal of Pediatric Obesity. 2009;4(4):266-273. doi:10.3109/17477160902811199

Subscales:
West F, Morawska A, Joughin K. The Lifestyle Behaviour Checklist: evaluation of the factor structure. Child: Care, Health and Development. 2010;36(4):508-515. doi:10.1111/j.1365-2214.2010.01074.x (\href{https://pubmed.ncbi.nlm.nih.gov/20337641/}{PubMed})
}
