% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_pss.R
\name{score_pss}
\alias{score_pss}
\title{score_pss: Scored data from the 10-item Perceived Stress Scale}
\usage{
score_pss(pss_data, score_base = TRUE, id)
}
\arguments{
\item{pss_data}{a data.frame all items for the Perceived Stress Scale following the naming conventions described above}

\item{score_base}{the smallest value assigned to a choice is 0 (i.e., range 0-3). Default = TRUE.}

\item{id}{(optional) name of participant ID column in bes_data. If included the output dataset will be matched by id, if not included the output dataset will be in the order of bes_data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}
}
\value{
A dataset with subscale scores for the Perceived Stress Scale
}
\description{
This function scores the 10-item Perceived Stress Scale and provides a total PSS score and subscale scores for: Perceived Helplessness and Perceived Self-efficacy
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention: 'pss#' or 'pss_#' where # is the question number (1-10)
3) Questions 1-10 must have the numeric value for the choices: 0 - Never, 1 - Almost, 2 - Sometimes, 3 - Fairly Often, 4 - Very Often

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the pss with IDs
pss_score_data <- score_pss(pss_data, id = 'ID')


\dontrun{
}

}
\references{
Cohen, S., Kamarck, T., and Mermelstein, R.(1983). A global measure of perceived stress. Journal of Health and Social Behavior, 24, 386-396. PMID: 6668417.
Cohen, S. and Williamson, G. Perceived Stress in a Probability Sample of the United States. Spacapan, S. and Oskamp, S. (Eds.) The Social Psychology of Health. Newbury Park, CA: Sage, 1988.


Two-factor subscale scoring comes from:
Taylor JM. Psychometric analysis of the Ten-Item Perceived Stress Scale. Psychol Assess. 2015 Mar;27(1):90-101. doi: 10.1037/a0038100. Epub 2014 Oct 27. PMID: 25346996.
}
