% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_scpf.R
\name{score_scpf}
\alias{score_scpf}
\title{score_scpf: Scored data from the Structure and Control in Parent Feeding questionnaire}
\usage{
score_scpf(scpf_data, base_zero = TRUE, id, extra_scale_cols = c())
}
\arguments{
\item{scpf_data}{a data.frame all items for the Structure and Control in Parent Feeding questionnaire following the naming conventions described above}

\item{base_zero}{(logical) TRUE indicates the smallest value assigned to a choice is 0. FALSE indicates indicates the smallest value assigned to a choice is 1. Default = TRUE.}

\item{id}{(optional) name of participant ID column in input data. If included, the output dataset will be matched by id, if not included the output dataset will be in the order of the input data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}

\item{extra_scale_cols}{a vector of character strings that begin with 'scpf' but are not scale items. Any columns in scpf_data that begin with 'scpf' but are not scale items must be included here. Default is empty vector.}
}
\value{
A dataset with subscale scores for the Structure and Control in Parent Feeding questionnaire
}
\description{
This function scores the Structure and Control in Parent Feeding questionnaire and provides subscale scores for: Limit Exposure, Consistent feeding routines, Restriction, Pressure to eat
}
\details{
For data to be scored correctly, the data must be prepared according to the following criteria: \cr
\itemize{
 \item{The data must include all individual questionnaire items}
 \item{The columns/variables must match the following naming convention: 'scpf#' or 'scpf_#' where # is the question number (1-34)}
 \item{All questionnaire responses must be a numeric value ranging from 0-4 (base_zero = TRUE) or 1-5 (base_zero = FALSE) where: }
 \itemize{
    \item{For base_zero = TRUE: 0 = Never; 1 = Rarely; 2 = Sometimes; 3 = Often; 4 = Always}
    \item{For base_zero = FALSE: 1 = Never; 2 = Rarely; 3 = Sometimes; 4 = Often; 5 = Always}
  }
 \item{Missing values must be coded as NA}
 \item{Values must not be reversed scored. This script will apply reverse scoring so all levels must be true to the scale described above}
}
Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the scpf with IDs
scpf_score_data <- score_scpf(scpf_data, id = 'ID', base_zero = TRUE)

\dontrun{
}


}
\references{
Savage, J.S., Rollins, B.Y., Kugler, K.C. et al. Development of a theory-based questionnaire to assess structure and control in parent feeding (SCPF). Int J Behav Nutr Phys Act 14, 9 (2017). https://doi.org/10.1186/s12966-017-0466-2
}
