% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_tfeq18.R
\name{score_tfeq18}
\alias{score_tfeq18}
\title{score_tfeq18: Scored data from the Three-Factor Eating Questionnaire-R18}
\usage{
score_tfeq18(tfeq_data, score_base = TRUE, id)
}
\arguments{
\item{tfeq_data}{a data.frame all items for the Three-Factor Eating Questionnaire-R18 following the naming conventions described above}

\item{score_base}{the smallest value assigned to a choice is 0 (i.e., range 0-3). Default = TRUE.}

\item{id}{(optional) name of participant ID column in bes_data. If included the output dataset will be matched by id, if not included the output dataset will be in the order of bes_data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}
}
\value{
A dataset with scores for the Three-Factor Eating Questionnaire-R18
}
\description{
This function scores the Three-Factor Eating Questionnaire-R18 and provides scores for cognitive restraint, uncontrolled eating, and emotional eating subscales
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention: 'tfeq#' where # is question number. 
3) Questions 1-13 must have the numeric value for the choices: 1, Definitely false | 2, Mostly false | 3, Mostly true | 4, Definitely true;
   Questions 14 must have the numeric value for the choices: 1, Only at meal times | 2, Sometimes between meals | 3, Often between meals | 4, Almost always
   Questions 15 must have the numeric value for the choices: 1, Almost never | 2, Seldom | 3, Usually | 4, Almost always
   Questions 16 must have the numeric value for the choices: 1, Unlikely | 2, Slightly likely | 3, Moderately likely | 4, Very likely
   Questions 17 must have the numeric value for the choices: 1,  Never | 2, Rarely | 3, Sometimes | 4, At least once a week
   Questions 18 must have the numeric values 1-8
Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the tfeq-r18 with IDs
tfeq_score_data <- score_tfeq18(tfeq_data, id = 'ID')


\dontrun{
}


}
\references{
The Three-Factor Eating Questionnaire-R18 Is Able to Distinguish among Different Eating Patterns in a General Population
}
